requirements.txt
PyYAML
sqlalchemy
redis